+++ package.json
@@ -19,6 +19,10 @@
     "react-native-web": "^0.11.4",
     "react-navigation": "^4.0.0",
     "react-navigation-stack": "^1.5.1",
+    "react-redux": "^7.1.1",
+    "redux": "^4.0.4",
+    "redux-thunk": "^2.3.0",
+    "remote-redux-devtools": "^0.5.16",
     "yarn": "^1.17.3"
   },
   "devDependencies": {
